Create useHoverLayer custom hook for New and Summary project maps
